fix : config 폴더 경로 변경
